src/tests/test_security_measures.py
```python
import unittest
from src import security_measures

class TestSecurityMeasures(unittest.TestCase):

    def setUp(self):
        self.security = security_measures.SecurityMeasures()

    def test_implement_security(self):
        result = self.security.implement_security()
        self.assertTrue(result, "Security measures implementation failed.")

    def test_data_encryption(self):
        test_data = "Test data for encryption"
        encrypted_data = self.security.encrypt_data(test_data)
        self.assertNotEqual(test_data, encrypted_data, "Data encryption failed.")

        decrypted_data = self.security.decrypt_data(encrypted_data)
        self.assertEqual(test_data, decrypted_data, "Data decryption failed.")

    def test_data_sanitization(self):
        test_data = "<script>alert('Test')</script>"
        sanitized_data = self.security.sanitize_data(test_data)
        self.assertNotIn("<script>", sanitized_data, "Data sanitization failed.")

    def test_secure_communication(self):
        result = self.security.secure_communication()
        self.assertTrue(result, "Secure communication setup failed.")

if __name__ == "__main__":
    unittest.main()
```